feat/#1: 스토어 키, 난독화 설정

+++ data/src/main/java/com/yeolsimee/moneysaving/data/api/model/SampleData.kt
@@ -4,7 +4,7 @@ data class SampleData(
     val totSellamnt: Double? = null,
     val returnValue: String? = null,
     val drwNoDate: String? = null,
-    val firstWinamnt: Int? = null,
+    val firstWinamnt: Double? = null,
     val drwNo: Double? = null,
     val drwtNo1: Double? = null,
     val drwtNo2: Double? = null,
@@ -12,7 +12,7 @@ data class SampleData(
     val drwtNo4: Double? = null,
     val drwtNo5: Double? = null,
     val drwtNo6: Double? = null,
-    val firstPrzwnerCo: Double? = null,
+    val firstPrzwnerCo: Long? = null,
     val bnusNo: Double? = null
 ) {
     fun isNullOrBlank(): Boolean {
